Reorganize csv read module
