+++ crates/polars-pipe/src/executors/sources/csv.rs
@@ -3,8 +3,9 @@ use std::path::PathBuf;
 
 use polars_core::export::arrow::Either;
 use polars_core::POOL;
-use polars_io::csv::read_impl::{BatchedCsvReaderMmap, BatchedCsvReaderRead};
-use polars_io::csv::{CsvEncoding, CsvParserOptions, CsvReader};
+use polars_io::csv::{
+    BatchedCsvReaderMmap, BatchedCsvReaderRead, CsvEncoding, CsvParserOptions, CsvReader,
+};
 use polars_plan::global::_set_n_rows_for_scan;
 use polars_plan::prelude::FileScanOptions;
 use polars_utils::iter::EnumerateIdxTrait;
